Move commands to dedicated folders
Move the content of the bootstrap folder to more adapted places
