+++ src/cli/commands/_common/error.ts
