+++ src/cli/commands/config/telemetry.ts
@@ -19,9 +19,9 @@
  */
 // Configure CLI settings
 
-import { loadState, saveState } from '../../lib/state-manager';
-import { info, success } from '../../ui';
-import { InvalidOptionError } from './common/error';
+import { loadState, saveState } from '../../../lib/state-manager';
+import { info, success } from '../../../ui';
+import { InvalidOptionError } from '../_common/error';
 
 export interface ConfigureTelemetryOptions {
   enabled?: boolean;
