+++ src/cli/commands/integrate/claude/hook-templates.ts
